Monitoring: Move QueryBrowser to dynamic SDK
Move the `QueryBrowser` component and `queryBrowserTheme` under
console-shared/ and expose from the SDK.
Move `QueryBrowserProps` and `FormatSeriesTitle` types to SDK.
Changed 2 `key` props to fix eslint `react/no-array-index-key` errors.
Rename `QueryBrowser_` to `QueryBrowserWrapped` and `Tooltip_` to
`TooltipWrapped` to fix eslint `no-underscore-dangle` errors.
Several other small code changes were automatically made by eslint.
