+++ frontend/packages/console-dynamic-plugin-sdk/src/api/dynamic-core-api.ts
@@ -29,6 +29,7 @@ import {
   TimestampProps,
   NamespaceBarProps,
   CodeEditorRef,
+  QueryBrowserProps,
 } from '../extensions/console-types';
 import { StatusPopupSectionProps, StatusPopupItemProps } from '../extensions/dashboard-types';
 
@@ -752,4 +753,43 @@ export const NamespaceBar: React.FC<NamespaceBarProps> = require('@console/inter
  * ```
  */
 export const ErrorBoundaryFallbackPage: React.FC<ErrorBoundaryFallbackProps> = require('@console/shared/src/components/error/fallbacks/ErrorBoundaryFallbackPage')
+  .default;
+
+/**
+ * A component that renders a graph of the results from a Prometheus PromQL query along with controls for interacting with the graph.
+ * @param {CustomDataSource} customDataSource - (optional) Base URL of an API endpoint that handles PromQL queries. If provided, this is used instead of the default API for fetching data.
+ * @param {number} defaultSamples - (optional) The default number of data samples plotted for each data series. If there are many data series, QueryBrowser might automatically pick a lower number of data samples than specified here.
+ * @param {number} defaultTimespan - (optional) The default timespan for the graph in milliseconds - defaults to 1,800,000 (30 minutes).
+ * @param {PrometheusLabels[][]} disabledSeries - (optional) Disable (don't display) data series with these exact label / value pairs.
+ * @param {boolean} disableZoom - (optional) Flag to disable the graph zoom controls.
+ * @param {PrometheusLabels} filterLabels - (optional) Optionally filter the returned data series to only those that match these label / value pairs.
+ * @param {number} fixedEndTime - (optional) Set the end time for the displayed time range rather than showing data up to the current time.
+ * @param {(labels: PrometheusLabels, i?: number) => string} formatSeriesTitle - (optional) Function that returns a string to use as the title for a single data series.
+ * @param {React.ComponentType<{}>} GraphLink - (optional) Component for rendering a link to another page (for example getting more information about this query).
+ * @param {boolean} hideControls - (optional) Flag to hide the graph controls for changing the graph timespan, and so on.
+ * @param {boolean} isStack - (optional) Flag to display a stacked graph instead of a line graph. If showStackedControl is set, it will still be possible for the user to switch to a line graph.
+ * @param {string} namespace - (optional) If provided, data is only returned for this namespace (only series that have this namespace label).
+ * @param {GraphOnZoom} onZoom - (optional) Callback called when the graph is zoomed.
+ * @param {number} pollInterval - (optional) If set, determines how often the graph is updated to show the latest data (in milliseconds).
+ * @param {string[]} queries - Array of PromQL queries to run and display the results in the graph.
+ * @param {boolean} showLegend - (optional) Flag to enable displaying a legend below the graph.
+ * @param {boolean} showStackedControl - Flag to enable displaying a graph control for switching between stacked graph mode and line graph mode.
+ * @param {number} timespan - (optional) The timespan that should be covered by the graph in milliseconds.
+ * @param {string} units - (optional) Units to display on the Y-axis and in the tooltip.
+
+
+ * @example
+ * ```tsx
+ * <QueryBrowser
+ *   defaultTimespan={15 * 60 * 1000}
+ *   namespace={namespace}
+ *   pollInterval={30 * 1000}
+ *   queries={[
+ *     'process_resident_memory_bytes{job="console"}',
+ *     'sum(irate(container_network_receive_bytes_total[6h:5m])) by (pod)',
+ *   ]}
+ * />
+ * ```
+ */
+export const QueryBrowser: React.FC<QueryBrowserProps> = require('@console/shared/src/components/query-browser')
   .default;
