+++ frontend/packages/console-dynamic-plugin-sdk/src/extensions/console-types.ts
@@ -670,3 +670,27 @@ export type ErrorBoundaryFallbackProps = {
   stack: string;
   title: string;
 };
+
+export type FormatSeriesTitle = (labels: PrometheusLabels, i?: number) => string;
+
+export type QueryBrowserProps = {
+  customDataSource?: CustomDataSource;
+  defaultSamples?: number;
+  defaultTimespan?: number;
+  disabledSeries?: PrometheusLabels[][];
+  disableZoom?: boolean;
+  filterLabels?: PrometheusLabels;
+  fixedEndTime?: number;
+  formatSeriesTitle?: FormatSeriesTitle;
+  GraphLink?: React.ComponentType<{}>;
+  hideControls?: boolean;
+  isStack?: boolean;
+  namespace?: string;
+  onZoom?: (from: number, to: number) => void;
+  pollInterval?: number;
+  queries: string[];
+  showLegend?: boolean;
+  showStackedControl?: boolean;
+  timespan?: number;
+  units?: string;
+};
