+++ frontend/packages/console-shared/src/components/query-browser/index.ts
@@ -1,0 +1,2 @@
+export * from './QueryBrowser';
+export * from './theme';
