+++ frontend/packages/console-shared/src/components/query-browser/theme.ts
