+++ frontend/public/locales/en/public.json
@@ -1088,15 +1088,6 @@
   "Access restricted.": "Access restricted.",
   "Failed to load metrics list.": "Failed to load metrics list.",
   "Clear query": "Clear query",
-  "graph timespan": "graph timespan",
-  "Reset zoom": "Reset zoom",
-  "Displaying with reduced resolution due to large dataset.": "Displaying with reduced resolution due to large dataset.",
-  "query browser chart": "query browser chart",
-  "Ungraphable results": "Ungraphable results",
-  "Query results include range vectors, which cannot be graphed. Try adding a function to transform the data.": "Query results include range vectors, which cannot be graphed. Try adding a function to transform the data.",
-  "Query result is a string, which cannot be graphed.": "Query result is a string, which cannot be graphed.",
-  "The resulting dataset is too large to graph.": "The resulting dataset is too large to graph.",
-  "Stacked": "Stacked",
   "Your default receiver will automatically receive all alerts from this cluster that are not caught by other receivers first.": "Your default receiver will automatically receive all alerts from this cluster that are not caught by other receivers first.",
   "The routing labels for this receiver are configured to capture critical alerts. Finish setting up this receiver by selecting a \"Receiver Type\" to choose a destination for these alerts. If this receiver is deleted, critical alerts will go to the default receiver instead.": "The routing labels for this receiver are configured to capture critical alerts. Finish setting up this receiver by selecting a \"Receiver Type\" to choose a destination for these alerts. If this receiver is deleted, critical alerts will go to the default receiver instead.",
   "The Watchdog alert fires constantly to confirm that your alerting stack is functioning correctly. This receiver is configured to prevent it from creating unnecessary notifications. You can edit this receiver if you plan to use the information that Watchdog provides, otherwise this receiver should remain in its current state with no set receiver type.": "The Watchdog alert fires constantly to confirm that your alerting stack is functioning correctly. This receiver is configured to prevent it from creating unnecessary notifications. You can edit this receiver if you plan to use the information that Watchdog provides, otherwise this receiver should remain in its current state with no set receiver type.",
@@ -1992,6 +1983,15 @@
   "Filter by unit": "Filter by unit",
   "Unit": "Unit",
   "Loading {{title}} status": "Loading {{title}} status",
+  "graph timespan": "graph timespan",
+  "Reset zoom": "Reset zoom",
+  "Displaying with reduced resolution due to large dataset.": "Displaying with reduced resolution due to large dataset.",
+  "query browser chart": "query browser chart",
+  "Ungraphable results": "Ungraphable results",
+  "Query results include range vectors, which cannot be graphed. Try adding a function to transform the data.": "Query results include range vectors, which cannot be graphed. Try adding a function to transform the data.",
+  "Query result is a string, which cannot be graphed.": "Query result is a string, which cannot be graphed.",
+  "The resulting dataset is too large to graph.": "The resulting dataset is too large to graph.",
+  "Stacked": "Stacked",
   "Disable": "Disable",
   "default": "default",
   "Disabled": "Disabled",
